+++ tensorcircuit/templates/blocks.py
@@ -4,15 +4,32 @@ shortcuts for measurement patterns on circuit
 # circuit in, circuit out
 # pylint: disable=invalid-name
 
-from typing import Any, Optional, Sequence, Tuple
+from functools import wraps
+from typing import Any, Callable, Optional, Sequence, Tuple
+
+import numpy as np
 
 from .graphs import Grid2DCoord
 from .. import gates as G
+from ..circuit import Circuit as Circ
 from ..cons import backend
 
 Circuit = Any  # we don't use the real circuit class as too many mypy complains emerge
 Tensor = Any
 Graph = Any
+
+
+def state_centric(f: Callable[..., Circuit]) -> Callable[..., Tensor]:
+    @wraps(f)
+    def wrapper(s: Tensor, *args: Any, **kws: Any) -> Tensor:
+        n = backend.sizen(s)
+        n = int(np.log2(n))
+        c = Circ(n, inputs=s)
+        c = f(c, *args, **kws)
+        s = c.state()
+        return s
+
+    return wrapper
 
 
 def Bell_pair_block(